Version ready for use.
Readfile function in DocumentReader class now opens the file for every other time opening a file is called for.

Added details to DocumentReader class like returning a message if a word is not found in a given file.

Also if one inputs a limit which is greater than the common_word list the limit is automatically set to the length of said list